refactor(tui): split at-cap mode_set.rs and view/authoring.rs
Both files sat at the 800-LOC budget cap with zero headroom.
Flat-sibling split mirrors the existing authoring/{mod,forms,mode_set} pattern.

- authoring/mode_set.rs: extract impl AuthoringMenuState handler block into
  mode_set_handlers.rs (799 → 238 LOC types; 583 LOC handlers)
- view/authoring.rs: extract six mode-set renderers into authoring_mode_set.rs
  (798 → 581 LOC dispatcher+lifecycle; 271 LOC mode-set renderers)
- also remove the stale authoring.rs that was never deleted in si6.3
